split into two modules, integration layer and core
