+++ lerna.json
@@ -14,6 +14,7 @@
     "libraries/functional-tests",
     "libraries/testbot",
     "libraries/botframework-streaming-extensions",
+    "libraries/botbuilder-streaming-extensions",
     "transcripts"
   ],
   "version": "independent",
