+++ libraries/botbuilder-streaming-extensions/.gitignore
@@ -1,0 +1,4 @@
+/lib
+/node_modules
+*.js.map
+/coverage/**
