+++ libraries/botbuilder-streaming-extensions/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "botframework-streaming-extensions",
+  "name": "botbuilder-streaming-extensions",
   "version": "0.0.1",
   "description": "",
   "main": "lib/index.js",
@@ -18,6 +18,12 @@
     "typescript": "3.1.1"
   },
   "dependencies": {
+    "@azure/ms-rest-js": "1.2.6",
+    "botbuilder": "~4.5.1",
+    "botbuilder-core": "~4.5.1",
+    "botframework-connector": "~4.5.1",
+    "botframework-schema": "~4.5.1",
+    "botframework-streaming-extensions": "0.0.1",
     "promise.prototype.finally": "^3.1.0",
     "uuid": "^3.3.2",
     "watershed": "^0.4.0"
