+++ libraries/botbuilder-streaming-extensions/src/index.ts
