+++ libraries/botframework-streaming-extensions/src/index.ts
@@ -7,8 +7,7 @@
  */
 export { ContentStream } from './ContentStream'; // Temporary for DirectLineJS integration
 export { HttpContent } from './HttpContentStream'; // Temporary for DirectLineJS integration
-export * from './Integration';
-export { IReceiveRequest } from './Interfaces/IReceiveRequest'; // Temporary for DirectLineJS integration
+export { IStreamingTransportServer, ISocket, IReceiveRequest } from './Interfaces' // Temporary for DirectLineJS integration
 export { NamedPipeClient, NamedPipeServer } from './NamedPipe';
 export { RequestHandler } from './RequestHandler';
 export { StreamingRequest } from './StreamingRequest';
