+++ package.json
@@ -36,7 +36,8 @@
     "typedoc": "^0.14.2",
     "typedoc-plugin-external-module-name": "^2.1.0",
     "typedoc-plugin-markdown": "^2.0.6",
-    "typescript": "^3.5.2"
+    "typescript": "^3.5.2",
+    "watershed": "^0.4.0"
   },
   "nyc": {
     "exclude": [
